okay so I added some stuff but my old changes dissapeared
what is going on I'm gonna try some stuff now
