+++ src/rougelikeLibrary/Enemy.java
@@ -1,0 +1,11 @@
+package rougelikeLibrary;
+
+public class Enemy extends LivingObject {
+
+	public Enemy(int speed, int health){
+		super(speed, health);
+		
+	}
+	
+	
+}
