+++ tests/rougelikeLibrary/LivingObjectTests.java
@@ -1,0 +1,5 @@
+package rougelikeLibrary;
+
+public class LivingObjectTests {
+
+}
